Extract AJKF scraper module
